printsh: add logging to file surpport
Adds support for logging to a file. This can be enabled in the config
file by setting the PRINTSH_LOG_TO_FILE to true and setting the path
to the log file in PRINTSH_LOG_FILE_PATH. The default path is set to
the current working directory and the file name is set to shell.log.