Add argument parser module
This commit adds a argument parser module that can be used to parse
arguments and options from a command line.

When the user enters a command the command needs to be parsed
so the commands can easily get the arguments and options without
needing to create a argument parser for each command

Example:
test -1 -2 -3

Will be parsed to:
["test", "-1", "-2", "-3"]

The function returns a vector of strings that contains the arguments
and options. The command can then use this to get the arguments and
options quickly and easily.

Signed-off-by: awesomelewis2007 <awesomelewis2007@users.noreply.github.com>